refactor: improve bot code quality and logging
- Extracted complex audio playback logic from `play` and `skip` commands into a new `AudioService`
- Injected `AudioService` into Sapphire container for dependency injection
- Abstracted `tslog` configuration to a shared `logger` in `@sirubot/utils`
- Updated both Bot and ShardManager to utilize the shared `logger` for unified output format
- Fixed inefficient logger instantiation in `SapphireInterfaceLogger`
