+++ apps/bot/src/modules/audio/commands/play.ts
@@ -152,7 +152,6 @@ export class PlayCommand extends Command {
 		await interaction.deferReply();
 
 		const voiceChannel = interaction.member.voice.channelId;
-
 		const query = interaction.options.getString('query', true);
 		const platform = (interaction.options.getString('platform') || 'ytsearch') as SearchPlatform;
 		const context = {
@@ -164,186 +163,35 @@ export class PlayCommand extends Command {
 			guildId: interaction.guildId
 		};
 
-		const player =
-			this.container.audio.getPlayer(interaction.guildId) ||
-			(await this.container.audio.createPlayer({
-				guildId: interaction.guildId,
-				voiceChannelId: voiceChannel,
-				textChannelId: interaction.channelId,
-				selfDeaf: true,
-				selfMute: false,
-				instaUpdateFiltersFix: true,
-				applyVolumeAsFilter: true
-			}));
+		const player = await this.container.audioService.getOrCreatePlayer(
+			interaction.guildId,
+			voiceChannel,
+			interaction.channelId
+		);
 
-		const searchRes = await player.search({ query, source: platform }, { id: interaction.user.id, username: interaction.user.username });
-		switch (searchRes.loadType) {
-			case 'error': {
-				throw new UserError({
-					identifier: 'play_search_error',
-					message: '🛠️  음악 검색 중 오류가 발생했어요. 잠시 후 다시 시도해 주세요.',
-					context
-				});
-			}
+		const searchRes = await this.container.audioService.search(
+			player,
+			query,
+			platform,
+			{ id: interaction.user.id, username: interaction.user.username },
+			context
+		);
 
-			case 'empty': {
-				throw new UserError({
-					identifier: 'play_search_empty',
-					message: '🔎  검색 결과가 없어요. 다른 검색어로 다시 시도해 주세요.',
-					context
-				});
-			}
-		}
+		const playerConnectedBeforeSearch = player.connected;
+		await this.container.audioService.connectPlayer(player, context);
 
-		const playerConnected = player.connected;
-
-		if (!playerConnected) {
-			try {
-				await player.connect();
-			} catch (error) {
-				this.container.logger.error(error);
-				throw new UserError({
-					message: '🛠️  음성 채널에 접속할 수 없어요. 음성 채널 권한을 확인해주세요.',
-					identifier: 'play_connect_error',
-					context
-				});
-			}
-		}
 		switch (searchRes.loadType) {
 			case 'playlist': {
-				const playlist = searchRes.playlist!;
-				if (playlist.selectedTrack) {
-					await player.queue.add(playlist.selectedTrack);
-
-					// 나머지 곡들 개수 계산
-					const remainingTracks = searchRes.tracks.filter((track) => track.info.identifier !== playlist.selectedTrack!.info.identifier);
-					const remainingTracksCount = remainingTracks.length;
-
-					if (remainingTracksCount > 0) {
-						// 나머지 곡들이 있으면 추가할지 묻기
-						const askReply = await interaction.editReply({
-							flags: MessageFlags.IsComponentsV2,
-							components: [
-								view.askPlaylistAdd({
-									playlist,
-									selectedTrack: playlist.selectedTrack as Track,
-									remainTracks: remainingTracks as Track[],
-									player
-								})
-							],
-							allowedMentions: { users: [], roles: [] }
-						});
-
-						const trackAdded = view.trackAdded({
-							track: playlist.selectedTrack as Track,
-							queued: player.queue.current !== null,
-							position: player.queue.tracks.length,
-							totalDuration: player.queue.tracks.reduce((acc, track) => acc + (track.info.duration ?? 0), 0)
-						});
-
-						// 버튼 상호작용을 위한 collector 설정
-						const buttonInteractionCollector = askReply.createMessageComponentCollector({
-							filter: (i) => i.user.id === interaction.user.id && i.customId.includes('playlist_'),
-							time: 30000,
-							componentType: ComponentType.Button
-						});
-
-						const handleEndAction = async () => {
-							await interaction.editReply({
-								flags: MessageFlags.IsComponentsV2,
-								components: [trackAdded],
-								allowedMentions: { users: [], roles: [] }
-							});
-						};
-
-						const handleButtonAction = async (collectorInteraction: ButtonInteraction<'cached'>) => {
-							await collectorInteraction.deferUpdate();
-							if (!player.connected) return;
-							buttonInteractionCollector.off('collect', handleButtonAction);
-							buttonInteractionCollector.off('end', handleEndAction);
-
-							if (collectorInteraction.customId === 'playlist_add_remaining') {
-								// 나머지 곡들을 대기열에 추가
-								await player.queue.add(remainingTracks);
-
-								await collectorInteraction.editReply({
-									flags: MessageFlags.IsComponentsV2,
-									components: [
-										view.playlistQueued({
-											playlist,
-											tracks: remainingTracks as Track[]
-										})
-									],
-									allowedMentions: { users: [], roles: [] }
-								});
-
-								return;
-							} else if (collectorInteraction.customId === 'playlist_skip_remaining') {
-								await collectorInteraction.editReply({
-									flags: MessageFlags.IsComponentsV2,
-									components: [trackAdded],
-									allowedMentions: { users: [], roles: [] }
-								});
-
-								return;
-							}
-						};
-
-						buttonInteractionCollector.once('collect', handleButtonAction);
-						buttonInteractionCollector.once('end', handleEndAction);
-					} else {
-						// 나머지 곡이 없으면 일반적인 재생 메시지 표시
-						await interaction.editReply({
-							flags: MessageFlags.IsComponentsV2,
-							components: [
-								view.trackAdded({
-									track: playlist.selectedTrack as Track,
-									queued: player.queue.current !== null,
-									position: player.queue.tracks.length,
-									totalDuration: player.queue.tracks.reduce((acc, track) => acc + (track.info.duration ?? 0), 0)
-								})
-							],
-							allowedMentions: { users: [], roles: [] }
-						});
-					}
-				} else {
-					// Playlist 추가
-					await player.queue.add(searchRes.tracks);
-
-					await interaction.editReply({
-						flags: MessageFlags.IsComponentsV2,
-						components: [
-							view.playlistQueued({
-								playlist,
-								tracks: searchRes.tracks as Track[]
-							})
-						],
-						allowedMentions: { users: [], roles: [] }
-					});
-				}
+				await this.container.audioService.handlePlaylistPlay(interaction, player, searchRes);
 				break;
 			}
 			case 'track':
 			case 'search': {
-				await player.queue.add(searchRes.tracks[0]);
-				await interaction.editReply({
-					flags: MessageFlags.IsComponentsV2,
-					components: [
-						view.trackAdded({
-							track: searchRes.tracks[0] as Track,
-							queued: player.queue.current !== null,
-							position: player.queue.tracks.length,
-							totalDuration: player.queue.tracks.reduce((acc, track) => acc + (track.info.duration ?? 0), 0)
-						})
-					],
-					allowedMentions: { users: [], roles: [] }
-				});
+				await this.container.audioService.handleTrackPlay(interaction, player, searchRes);
 				break;
 			}
 		}
 
-		if (!player.playing) {
-			await player.play(playerConnected ? { paused: false } : undefined);
-		}
+		await this.container.audioService.ensurePlayback(player, playerConnectedBeforeSearch);
 	}
 }
