+++ apps/bot/src/types/global.d.ts
@@ -39,6 +39,7 @@ declare module '@sapphire/pieces' {
 		playerNotifier: PlayerNotifier;
 		guildService: GuildService;
 		trackService: TrackService;
+		audioService: import('../services/audioService.ts').AudioService;
 		shardClient?: import('@sirubot/shardclient').ShardClient;
 	}
 }
