+++ packages/utils/src/index.ts
@@ -19,3 +19,4 @@ export * from './youtube.js';
 export * from './array.js';
 export * from './embed.js';
 export * from './memoryCache.js';
+export * from './logger.js';
